+++ CHANGES.rst
@@ -5,6 +5,11 @@ Changelog
 0.2 (unreleased)
 ----------------
 
+- When generating link to sub faceted, make sure every default values are
+  correctly initialized, especially value of the Collection widget for which
+  the id could be different than current faceted Collection widget id
+  [gbastien]
+
 - If we use functionnality of sub folders where faceted navigation is enabled,
   and the default collection is on one of these subfolders, do not redirect to
   this default collection if we use the 'folder_contents' of the root folder or